fix(types): correct two wrong return-type annotations
example_2v2_setup()/example_3v3_setup() were annotated to return
SwarmTeam objects but actually return the dict[str, AgentRole] model
configs meant to be passed INTO create_team() -- every real call site
(CLI, web routes, tests) already treats them as dicts. get_pool_stats()
was annotated Dict[str, int] but its 'utilization' field is a float
ratio. Both were propagating false type errors into unrelated call
sites; no behavior change.

+++ ollama_arena/storage/sqlite/_conn.py
@@ -228,7 +228,7 @@ def close_all_connections():
         _connection_pool.clear()
 
 
-def get_pool_stats() -> Dict[str, int]:
+def get_pool_stats() -> Dict[str, float]:
     """Get statistics about the connection pool."""
     with _pool_lock:
         return {
